funcion eliminar en pagos

+++ Sistema/pagos/Gestor_tbl_Pago.php
@@ -112,7 +112,7 @@ if ($datos=$sql->fetch_object()) {
 }
 $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol and ID_Objeto=6");
 if ($datos=$sql->fetch_object()) { 
-        $output['data'] .= "<td><a class='boton-eliminar' href='Delete_Pago.php?ID_de_pago=" . $row['ID_de_pago'] . "'><i class='zmdi zmdi-delete'></i></a></td>";
+        $output['data'] .= "<td><a onclick='return confirmar()' class='boton-eliminar' href='Delete_Pago.php?ID_de_pago=" . $row['ID_de_pago'] . "'><i class='zmdi zmdi-delete'></i></a></td>";
 }
         $output['data'] .= '</tr>';
     }
